requirements.txt
# Main Framework
streamlit

# Database
sqlalchemy
sqlalchemy-libsql

# Data & Plotting
pandas
plotlyalembic
alembic
python-dotenv
